first draft
Supports many plot types and has flexible settings for creating menus in the user interface, with an option to show or hide them. Published visualizations can be embedded into a website, when menus are hidden, the app has the capacity of listening to calls from buttons in the main website so as to react to them. This version uses a local copy of the warehouse. Main options are directly defined in the app.R file.